+++ pykafka/__init__.py
@@ -8,7 +8,7 @@ from .client import KafkaClient
 from .balancedconsumer import BalancedConsumer
 from .managedbalancedconsumer import ManagedBalancedConsumer
 
-__version__ = '2.3.1'
+__version__ = '2.3.2-dev'
 
 
 __all__ = ["Broker", "SimpleConsumer", "Cluster", "Partition", "Producer",